transaction_service/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-filter==23.1
mysqlclient==2.1.1
requests==2.28.2
python-dotenv==1.0.0